Add custom CSS to fix image distortion
Use custom CSS to preserve the aspect ratio of cell output images. Sets height to auto to fix a problem with image distortion (caused by fixed height and width) that seems to have been introduced with Jupyter Book 1.0.0. (Note that this fixes the problem only when building the HTML, not when building the PDF - for building the PDF, consider switching to Jupyter Book 0.15.1, which does not distort the images.)